cityos_mqttsubsc/requirements.txt
elasticsearch
fastapi
uvicorn
python-multipart
sqlalchemy
jinja2
starlette
aiofiles
mysqlclient==2.1.1
requests
datetime
chardet
paho-mqtt
python-dotenv